requirements.txt
apache-libcloud==3.3.0
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
chicken==0.1.0
cloudpickle==1.6.0
cycler==0.10.0
egg==0.2.0
future==0.18.2
futures==3.1.1
idna==2.10
imutils==0.5.3
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.1.0
pkg-resources==0.0.0
py-tes==0.4.1
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.6.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
